+++ docs/changes.rst
@@ -12,6 +12,7 @@ BREAKING CHANGE: UI freshup
 --------------------
 
 - Prevent execution of malicious code entered as custom measure or training notes.
+- Translation fixes for form validation messages.
 
 12.0.16 (2021-11-11)
 --------------------